Complete UART Response handling rework
Now we work on a character basis.

The burden of implementation of buffering incoming characters lies
with the platform-specific code.

Response handling was completely reworked and is now stable even with
unrelated messages on the serial line - these will be skipped for now
but may receive a callback mechanism later, in case they are
relevant to the library user.